// nodejs/addcopyright/Source/app/shared/dto/dashboard/column-data-info-dto.ts
/* **********************************************************************
 * Copyright 2017 VMware, Inc.  All rights reserved. VMware Confidential
 * *********************************************************************/

export interface ColumnDataInfoDto {
    columnHeader: string;
    columnValue: string;
    columnType: string;
}